+++ src/main/java/org/cloudwarp/probablychests/world/feature/PCFeatures.java
@@ -6,7 +6,6 @@ import net.minecraft.util.math.floatprovider.UniformFloatProvider;
 import net.minecraft.util.registry.Registry;
 import net.minecraft.util.registry.RegistryEntry;
 import net.minecraft.world.Heightmap;
-import net.minecraft.world.gen.YOffset;
 import net.minecraft.world.gen.blockpredicate.BlockPredicate;
 import net.minecraft.world.gen.feature.*;
 import net.minecraft.world.gen.placementmodifier.*;
@@ -83,8 +82,9 @@ public class PCFeatures {
 
 	public static void init () {
 		Config config = ProbablyChests.config;
-		float chestRarity = config.getChestFrequency();
-		float potRarity = config.getPotFrequency();
+		float chestRarity = config.getChestSpawnChance();
+		float potRarity = config.getPotSpawnChance();
+		float surfaceChestRarity = config.getSurfaceChestSpawnChance();
 		//-----------------------------------
 		LUSH_CHEST_PLACED = PlacedFeatures.register("lush_chest_placed",
 				LUSH_CHEST, PCRarityFilterPlacementModifier.of(chestRarity),  SquarePlacementModifier.of(),
@@ -92,8 +92,8 @@ public class PCFeatures {
 						20, Heightmap.Type.WORLD_SURFACE_WG, 300),
 				BiomePlacementModifier.of());
 		LUSH_CHEST_PLACED_SURFACE = PlacedFeatures.register("lush_chest_placed_surface",
-				LUSH_CHEST_SURFACE, PCRarityFilterPlacementModifier.of(chestRarity), SquarePlacementModifier.of(),
-				PlacedFeatures.MOTION_BLOCKING_HEIGHTMAP,
+				LUSH_CHEST_SURFACE, SquarePlacementModifier.of(), PlacedFeatures.MOTION_BLOCKING_HEIGHTMAP,
+				PCRarityFilterPlacementModifier.of(surfaceChestRarity), PCSolidGroundPlacementModifier.of(BlockPredicate.hasSturdyFace(Direction.UP)),
 				BiomePlacementModifier.of());
 		//------------------------------------------------
 		NORMAL_CHEST_PLACED = PlacedFeatures.register("normal_chest_placed",
@@ -102,8 +102,8 @@ public class PCFeatures {
 						20, Heightmap.Type.WORLD_SURFACE_WG, 300),
 				BiomePlacementModifier.of());
 		NORMAL_CHEST_PLACED_SURFACE = PlacedFeatures.register("normal_chest_placed_surface",
-				NORMAL_CHEST_SURFACE, PCRarityFilterPlacementModifier.of(chestRarity), SquarePlacementModifier.of(),
-				PlacedFeatures.MOTION_BLOCKING_HEIGHTMAP,
+				NORMAL_CHEST_SURFACE,  SquarePlacementModifier.of(), PlacedFeatures.MOTION_BLOCKING_HEIGHTMAP,
+				PCRarityFilterPlacementModifier.of(surfaceChestRarity), PCSolidGroundPlacementModifier.of(BlockPredicate.hasSturdyFace(Direction.UP)),
 				BiomePlacementModifier.of());
 		//------------------------------------------------
 		ROCKY_CHEST_PLACED = PlacedFeatures.register("rocky_chest_placed",
@@ -112,8 +112,8 @@ public class PCFeatures {
 						20, Heightmap.Type.WORLD_SURFACE_WG, 300),
 				BiomePlacementModifier.of());
 		ROCKY_CHEST_PLACED_SURFACE = PlacedFeatures.register("rocky_chest_placed_surface",
-				ROCKY_CHEST_SURFACE, PCRarityFilterPlacementModifier.of(chestRarity), SquarePlacementModifier.of(),
-				PlacedFeatures.MOTION_BLOCKING_HEIGHTMAP,
+				ROCKY_CHEST_SURFACE,  SquarePlacementModifier.of(), PlacedFeatures.MOTION_BLOCKING_HEIGHTMAP,
+				PCRarityFilterPlacementModifier.of(surfaceChestRarity), PCSolidGroundPlacementModifier.of(BlockPredicate.hasSturdyFace(Direction.UP)),
 				BiomePlacementModifier.of());
 		//---------------------------------------------------------
 		STONE_CHEST_PLACED = PlacedFeatures.register("stone_chest_placed",