+++ changelog.md
@@ -18,3 +18,16 @@ All notable changes to this project will be documented in this file.
 - Biome specific spawning
 - Spawn rate config
 - Made Pots silk touchable
+
+## [0.3.0] - 4/12/2022
+### Added
+- Stone Chest
+- Gold Chest
+- Mimics
+- Mimic keys
+- Mimic Key Fragments
+### Changed
+- New chest spawn mechanics
+- Config options for rarity are now a percent float value
+- New snazzy icon
+- The proper chest types now mine faster with right tools